Make backend runnable with only Postgres; fix CSRF flow and live-DB bugs
Run the real full-stack app on Windows without Docker/Redis/S3/AI keys, and
fix the bugs that surfaced once it ran against a live Postgres (Neon).

Graceful degradation (only DATABASE_URL is required to boot):
- env: make Redis/S3/AI/SMTP optional with capability flags + dev defaults;
  parse booleans properly (z.coerce.boolean() treated "false" as true, which
  wrongly set Secure on cookies and broke them over http://localhost).
- storage: local-disk driver + /media route when no S3.
- redis: in-memory cache + rate-limit fallback when no REDIS_URL.
- queue: run the image pipeline inline (no separate worker) without Redis;
  extract the processor into jobs/imageProcessor.ts.
- ai/embeddings/chat/critique: non-AI fallbacks (plain metadata, deterministic
  embeddings, stats-based chat) so uploads/search/chat work without keys.
- mailer: log instead of send when SMTP is unconfigured.
- logger: fall back to JSON if pino-pretty isn't installed.
- startup banner shows which capabilities are on vs. degraded.

Live-DB fixes:
- knowledge seed: cast bound param to the KnowledgeKind enum (was inserting
  text into an enum column and silently failing every startup).
- pgvector calls: Prisma binds JS numbers as bigint, so search_photos /
  retrieve_knowledge / similar_photos / similar_by_palette / nearby_photos /
  find_duplicate_photos all threw "function does not exist" — add ::int/::real
  casts at every call site (search + chat + recommendations were broken).

CSRF flow (protection kept enabled end to end):
- add GET /api/auth/csrf that sets the cookie and returns the token in the body.
- frontend ensureCsrf() obtains the token before any mutating request instead
  of assuming the cookie already exists, so a fresh page load no longer races
  ahead of the cookie and gets a 403. Applied to api(), the upload XHR, and the
  chat stream. Verified in a real headless Chrome (warm + cold-start paths).

Docs: README quickstart + docs/RUN_ON_WINDOWS.md; .env.example rewritten so
optional services are commented out (copy + set DATABASE_URL = working run).

+++ server/src/modules/search/search.service.ts
@@ -22,7 +22,7 @@ async function runSearch(query: string, k: number) {
   const [vector] = await embed(query);
 
   const rows = await prisma.$queryRawUnsafe<SearchRow[]>(
-    `SELECT * FROM search_photos($1::vector, $2, $3)`,
+    `SELECT * FROM search_photos($1::vector, $2, $3::int)`,
     toVectorLiteral(vector),
     query,
     k
@@ -56,7 +56,7 @@ async function runSearch(query: string, k: number) {
 
 export async function similarPhotos(photoId: string, k = 6) {
   const rows = await prisma.$queryRawUnsafe<{ photo_id: string; score: number }[]>(
-    `SELECT * FROM similar_photos($1::uuid, $2)`,
+    `SELECT * FROM similar_photos($1::uuid, $2::int)`,
     photoId,
     k
   );
